Added hecto discrete grid support to RoadSegment
Added hecto_key to RoadSegment model and configured an index on it.

Exposed new endpoint  road_segments/:factor/:key  to enable discrete
grid querying of segment data. The factor is 1000, 10000, or 100000;
representing a multiplier on lat/lng. The key is a pair of 30-bit
base64 encoded strings;lat and lng, respectively.

Enabled NewRelic monitoring and MemCachier caching via dalli gem.